+++ packages/aware-components/src/components/index.ts
@@ -14,6 +14,7 @@ export { Label } from './Label';
 export { Main } from './Main';
 export { Nav } from './Nav';
 export { Optgroup } from './Optgroup';
+export { P } from './P';
 export { Section } from './Section';
 export { Select } from './Select';
 export { Textarea } from './Textarea';